Move wasmer_compiler/{address_map, function, module, sourceloc, trap, unwind}.rs to wasmer_types
