+++ lib/compiler-cranelift/src/compiler.rs
@@ -21,13 +21,16 @@ use gimli::write::{Address, EhFrame, FrameTable};
 #[cfg(feature = "rayon")]
 use rayon::prelude::{IntoParallelRefIterator, ParallelIterator};
 use std::sync::Arc;
-use wasmer_compiler::{CallingConvention, ModuleTranslationState, Target, TrapInformation};
+use wasmer_compiler::{CallingConvention, ModuleTranslationState, Target};
 use wasmer_compiler::{
-    Compilation, CompileModuleInfo, CompiledFunction, CompiledFunctionFrameInfo,
-    CompiledFunctionUnwindInfo, Compiler, Dwarf, FunctionBinaryReader, FunctionBody,
-    FunctionBodyData, MiddlewareBinaryReader, ModuleMiddleware, ModuleMiddlewareChain,
+    Compiler, FunctionBinaryReader, FunctionBodyData, MiddlewareBinaryReader, ModuleMiddleware,
+    ModuleMiddlewareChain,
 };
 use wasmer_types::entity::{EntityRef, PrimaryMap};
+use wasmer_types::{
+    Compilation, CompileModuleInfo, CompiledFunction, CompiledFunctionFrameInfo,
+    CompiledFunctionUnwindInfo, Dwarf, FunctionBody, TrapInformation,
+};
 use wasmer_types::{CompileError, FunctionIndex, LocalFunctionIndex, ModuleInfo, SignatureIndex};
 use wasmer_types::{Relocation, RelocationTarget, SectionIndex, TrapCode};
