+++ lib/compiler-cranelift/src/translator/unwind.rs
@@ -5,8 +5,8 @@
 use cranelift_codegen::isa::unwind::{systemv::UnwindInfo as DwarfFDE, UnwindInfo};
 use cranelift_codegen::print_errors::pretty_error;
 use cranelift_codegen::{isa, Context};
-use wasmer_compiler::CompiledFunctionUnwindInfo;
 use wasmer_types::CompileError;
+use wasmer_types::CompiledFunctionUnwindInfo;
 
 /// Cranelift specific unwind info
 pub(crate) enum CraneliftUnwindInfo {
