+++ lib/compiler-llvm/src/object_file.rs
@@ -4,15 +4,11 @@ use std::collections::{HashMap, HashSet};
 use std::convert::TryInto;
 use std::num::TryFromIntError;
 
-use wasmer_compiler::{
-    CompiledFunctionFrameInfo, CustomSections, FunctionAddressMap, FunctionBody,
-    InstructionAddressMap, SourceLoc,
-};
 use wasmer_types::entity::PrimaryMap;
-use wasmer_types::CompileError;
 use wasmer_types::{
-    CustomSection, CustomSectionProtection, Relocation, RelocationKind, RelocationTarget,
-    SectionBody, SectionIndex,
+    CompileError, CompiledFunctionFrameInfo, CustomSection, CustomSectionProtection,
+    CustomSections, FunctionAddressMap, FunctionBody, InstructionAddressMap, Relocation,
+    RelocationKind, RelocationTarget, SectionBody, SectionIndex, SourceLoc,
 };
 use wasmer_vm::libcalls::LibCall;
 
@@ -25,7 +21,7 @@ fn map_object_err(error: object::read::Error) -> CompileError {
 }
 
 pub struct CompiledFunction {
-    pub compiled_function: wasmer_compiler::CompiledFunction,
+    pub compiled_function: wasmer_types::CompiledFunction,
     pub custom_sections: CustomSections,
     pub eh_frame_section_indices: Vec<SectionIndex>,
 }
@@ -333,7 +329,7 @@ where
     };
 
     Ok(CompiledFunction {
-        compiled_function: wasmer_compiler::CompiledFunction {
+        compiled_function: wasmer_types::CompiledFunction {
             body: function_body,
             relocations: relocations
                 .remove_entry(&root_section_index)
