+++ lib/compiler-singlepass/src/emitter_arm64.rs
@@ -9,9 +9,9 @@ use dynasmrt::{
     aarch64::Aarch64Relocation, AssemblyOffset, DynamicLabel, DynasmApi, DynasmLabelApi,
     VecAssembler,
 };
-use wasmer_compiler::{CallingConvention, FunctionBody};
+use wasmer_compiler::CallingConvention;
 use wasmer_types::{CustomSection, CustomSectionProtection, SectionBody};
-use wasmer_types::{FunctionIndex, FunctionType, Type, VMOffsets};
+use wasmer_types::{FunctionBody, FunctionIndex, FunctionType, Type, VMOffsets};
 
 type Assembler = VecAssembler<Aarch64Relocation>;
