+++ lib/compiler-singlepass/src/machine.rs
@@ -8,11 +8,9 @@ use std::collections::BTreeMap;
 use std::fmt::Debug;
 pub use wasmer_compiler::wasmparser::MemoryImmediate;
 use wasmer_compiler::wasmparser::Type as WpType;
-use wasmer_compiler::{
-    Architecture, CallingConvention, CpuFeature, FunctionBody, InstructionAddressMap, Target,
-    TrapInformation,
-};
+use wasmer_compiler::{Architecture, CallingConvention, CpuFeature, Target};
 use wasmer_types::{CustomSection, Relocation, RelocationTarget};
+use wasmer_types::{FunctionBody, InstructionAddressMap, TrapInformation};
 use wasmer_types::{FunctionIndex, FunctionType, TrapCode, VMOffsets};
 
 pub type Label = DynamicLabel;
