+++ lib/compiler-singlepass/src/machine_arm64.rs
@@ -10,10 +10,9 @@ use dynasmrt::{aarch64::Aarch64Relocation, VecAssembler};
 #[cfg(feature = "unwind")]
 use gimli::{write::CallFrameInstruction, AArch64};
 use wasmer_compiler::wasmparser::Type as WpType;
-use wasmer_compiler::{
-    CallingConvention, FunctionBody, InstructionAddressMap, SourceLoc, TrapInformation,
-};
+use wasmer_compiler::CallingConvention;
 use wasmer_types::{CustomSection, Relocation, RelocationKind, RelocationTarget};
+use wasmer_types::{FunctionBody, InstructionAddressMap, SourceLoc, TrapInformation};
 use wasmer_types::{FunctionIndex, FunctionType, TrapCode, VMOffsets};
 
 type Assembler = VecAssembler<Aarch64Relocation>;
