+++ lib/compiler-singlepass/src/machine_x64.rs
@@ -13,13 +13,12 @@ use dynasmrt::{x64::X64Relocation, DynasmError, VecAssembler};
 use gimli::{write::CallFrameInstruction, X86_64};
 use std::ops::{Deref, DerefMut};
 use wasmer_compiler::wasmparser::Type as WpType;
-use wasmer_compiler::{
-    CallingConvention, CpuFeature, FunctionBody, InstructionAddressMap, SourceLoc, TrapInformation,
-};
+use wasmer_compiler::{CallingConvention, CpuFeature};
 use wasmer_types::{
     CustomSection, CustomSectionProtection, Relocation, RelocationKind, RelocationTarget,
     SectionBody,
 };
+use wasmer_types::{FunctionBody, InstructionAddressMap, SourceLoc, TrapInformation};
 use wasmer_types::{FunctionIndex, FunctionType, TrapCode, Type, VMOffsets};
 
 type Assembler = VecAssembler<X64Relocation>;
