+++ lib/types/src/compilation/mod.rs
@@ -1,4 +1,10 @@
 //! Types for compilation.
 
+pub mod address_map;
+pub mod function;
+pub mod module;
 pub mod relocation;
 pub mod section;
+pub mod sourceloc;
+pub mod trap;
+pub mod unwind;
